feat(storage): add FeatureAnnotationStorage interface
This commit introduces the `FeatureAnnotationStorage` interface, which
defines the contract for interacting with feature annotation data. This
interface outlines essential operations such as creating, retrieving,
updating, and deleting feature annotations, as well as managing
associated tags and listing annotations by PubMed ID or DOI. This
abstraction will enable easier testing and allow for different storage
backends to be implemented in the future.