feat(storage): add AddTags method to MemoryStorage
This change introduces the `AddTags` method to the `MemoryStorage`
struct. This method allows for the efficient addition of multiple tag
properties to a feature annotation in a single operation. Previously,
tags had to be added one by one using the `AddTag` method, which was
less efficient for bulk operations. The new method also includes checks
to prevent duplicate tags from being added to the same annotation.