chore(go.mod): update Kubernetes client libraries to v0.33.3
This update brings the Kubernetes client libraries to the latest stable
version, v0.33.3. The removal of `imdario/mergo` is due to it being an indirect
dependency that is no longer required by the direct dependencies. The
addition of `gopkg.in/evanphx/json-patch.v4` is a new indirect
dependency required by the updated Kubernetes libraries.