refactor(load_hypothetical_gene_products.go): extract and rename functions for better readability and reusability
The `processAnnotation` function has been refactored to improve its
readability and reusability by extracting and renaming several anonymous
functions into named functions.Specifically, the logic for handling a
missing annotation and the logic for checking and adding a hypothetical
product tag have been extracted into `handleAnnotationNotFound` and
`checkAndHandleProductTag` respectively. The `addProductTag` function
was also extracted to encapsulate the action of adding a product tag.
This change makes the `processAnnotation` function more concise and
easier to understand, while the extracted functions can be potentially
reused elsewhere if needed. The use of `F.Curry2` in
`checkAndHandleProductTag` also makes the function signature cleaner.